+++ src/designs.js
@@ -78,7 +78,7 @@ function intToHex(intValue) {
     return hexValue;
 }
 
-export function getDesignTokens({ vtcBackground, webConfig }, { themeSettings, setThemeSettings }, customMode, mode, use_custom_theme = false, theme_background = null, theme_main = null, darken_ratio = null, font_size = "regular") {
+export function getDesignTokens({ vtcBackground, customBackground, webConfig }, { themeSettings, setThemeSettings }, customMode, mode, use_custom_theme = false, theme_background = null, theme_main = null, darken_ratio = null, font_size = "regular") {
     if (use_custom_theme === true) {
         if (vars.userLevel < 2)
             use_custom_theme = false;
@@ -108,7 +108,7 @@ export function getDesignTokens({ vtcBackground, webConfig }, { themeSettings, s
             return; // we know there'll be a re-render
         }
     }
-    if (use_custom_theme === "custombg" && (vars.userLevel === -1 || vars.userLevel >= 3) && vars.dhcustombg !== "") {
+    if (use_custom_theme === "custombg" && (vars.userLevel === -1 || vars.userLevel >= 3) && customBackground !== "") {
         if (darken_ratio === null) darken_ratio = 0.4;
         if (mode === "dark") {
             if (theme_background === null) theme_background = "#212529";
@@ -119,8 +119,8 @@ export function getDesignTokens({ vtcBackground, webConfig }, { themeSettings, s
         }
         theme_background = theme_background.substring(0, 7) + intToHex(darken_ratio * 100);
         theme_main = theme_main.substring(0, 7) + intToHex(darken_ratio * 100);
-        if (setThemeSettings !== undefined && themeSettings.bg_image !== vars.dhcustombg) { // ensure called from <App>
-            setThemeSettings(prev_settings => ({ ...prev_settings, bg_image: vars.dhcustombg }));
+        if (setThemeSettings !== undefined && themeSettings.bg_image !== customBackground) { // ensure called from <App>
+            setThemeSettings(prev_settings => ({ ...prev_settings, bg_image: customBackground }));
             return; // we know there'll be a re-render
         }
     }